refactor(as7262): move Configure(...) to vControlRegister.go
